requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
google-generativeai
duckduckgo_search